# frontend/client-code-gen/fam-openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Forest Access Management - FAM - API",
    "description" : "Forest Access Management API. Defines who has access to which applications, and the roles they operate under once access is granted.\n\nCovers both the app-access-control and admin-management surfaces that upstream FAM served from two separate APIs.",
    "contact" : {
      "name" : "Team Heartwood",
      "url" : "https://apps.nrs.gov.bc.ca/int/confluence/display/FSAST1/Team+Heartwood",
      "email" : "SIBIFSAF@Victoria1.gov.bc.ca"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Relative to the host serving this document"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
    "/users/users-information" : {
      "put" : {
        "tags" : [ "FAM User" ],
        "summary" : "Refresh stored user details from IDIM",
        "description" : "Guarded by the X-API-Key shared secret, not a bearer token.",
        "operationId" : "update_user_information_from_idim_source",
        "parameters" : [ {
          "name" : "X-API-Key",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        }, {
          "name" : "perPage",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 100
          }
        }, {
          "name" : "usePagination",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FamUserUpdateResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "apiKey" : [ ]
        } ]
      }
    },
...
    "/identity-lookup/users/idir/search" : {
      "get" : {
        "tags" : [ "IDIR/BCeID Proxy" ],
        "summary" : "Search IDIR users",
        "description" : "Search for IDIR users.",
        "operationId" : "search_idir_users",
        "parameters" : [ {
          "name" : "firstName",
          "in" : "query",
          "description" : "IDIR first name search value (min 2 chars)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "IDIR first name search value (min 2 chars)",
            "maxLength" : 50,
            "minLength" : 0
          }
        }, {
          "name" : "lastName",
          "in" : "query",
          "description" : "IDIR last name search value (min 2 chars)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "IDIR last name search value (min 2 chars)",
            "maxLength" : 50,
            "minLength" : 0
          }
        }, {
          "name" : "userId",
          "in" : "query",
          "description" : "IDIR user id search value (min 2 chars)",
          "required" : false,
          "schema" : {
            "type" : "string",
            "description" : "IDIR user id search value (min 2 chars)",
            "maxLength" : 20,
            "minLength" : 0
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "Number of records to return",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Number of records to return",
            "maximum" : 500,
            "minimum" : 10
          }
        }, {
          "name" : "applicationId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IdimProxyIdirUsersSearchResultDto"
                }
              }
            }
          }
        }
      }
    },
    "/identity-lookup/idir" : {
      "get" : {
        "tags" : [ "IDIR/BCeID Proxy" ],
        "summary" : "Lookup IDIR user",
        "description" : "Lookup an IDIR user by user ID.",
        "operationId" : "idir_lookup",
        "parameters" : [ {
          "name" : "userId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "maxLength" : 20,
            "minLength" : 0
          }
        }, {
          "name" : "applicationId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IdimProxyIdirInfoDto"
                }
              }
            }
          }
        }
      }
    },
    "/identity-lookup/bceid" : {
      "get" : {
        "tags" : [ "IDIR/BCeID Proxy" ],
        "summary" : "Lookup BCEID user",
        "description" : "Lookup a BCeID Business user by user ID.",
        "operationId" : "bceid_lookup",
        "parameters" : [ {
          "name" : "userId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "maxLength" : 20,
            "minLength" : 0
          }
        }, {
          "name" : "applicationId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IdimProxyBceidInfoDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "FamUserUpdateEntryDto" : {
        "properties" : {
          "user_id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "user_name" : {
            "type" : "string"
          },
          "user_type" : {
            "type" : "string"
          },
          "user_guid" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        }
      },
      "FamUserUpdateResponse" : {
        "properties" : {
          "total_db_users_count" : {
            "type" : "integer",
            "format" : "int64"
          },
          "current_page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "users_count_on_page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "run_on" : {
            "type" : "string",
            "format" : "date-time"
          },
          "elapsed" : {
            "type" : "string"
          },
          "success_user_update_list" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FamUserUpdateEntryDto"
            }
          },
          "failed_user_update_list" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FamUserUpdateEntryDto"
            }
          },
          "ignored_user_update_list" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FamUserUpdateEntryDto"
            }
          },
          "mismatch_user_update_list" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FamUserUpdateEntryDto"
            }
          }
        }
      },
...
      "IdimProxyIdirUserSearchItemDto" : {
        "properties" : {
          "userId" : {
            "type" : "string"
          },
          "guid" : {
            "type" : "string"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        },
        "required" : [ "email", "firstName", "guid", "lastName", "userId" ]
      },
      "IdimProxyIdirUsersSearchResultDto" : {
        "properties" : {
          "totalItems" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IdimProxyIdirUserSearchItemDto"
            }
          }
        },
        "required" : [ "items", "pageSize", "totalItems" ]
      },
      "IdimProxyIdirInfoDto" : {
        "properties" : {
          "found" : {
            "type" : "boolean"
          },
          "userId" : {
            "type" : "string"
          },
          "guid" : {
            "type" : "string"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        },
        "required" : [ "found", "userId" ]
      },
      "IdimProxyBceidInfoDto" : {
        "properties" : {
          "found" : {
            "type" : "boolean"
          },
          "userId" : {
            "type" : "string"
          },
          "guid" : {
            "type" : "string"
          },
          "businessGuid" : {
            "type" : "string"
          },
          "businessLegalName" : {
            "type" : "string"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          }
        },
        "required" : [ "found", "userId" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "Access token issued by BC Gov SSO (Keycloak).",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "apiKey" : {
        "type" : "apiKey",
        "description" : "Shared secret for the scheduled user-information refresh.",
        "name" : "X-API-Key",
        "in" : "header"
      }
    }
  }
}